+++ lens-labels/Changelog.md
@@ -4,6 +4,8 @@
 - Improve readability of `HasLens` instances. (#118)
 - Remove support for `ghc-7.10`. (#136)
 - Use a `.cabal` file that's auto-generated from `hpack`. (#138)
+- Add `Lens.Labels.Unwrapped`, allowing easier interoperation with
+  other lens libraries. (#157)
 
 ## v0.1.0.2
 - Bump the dependency on `base` to support `ghc-8.2.1`.